Test Program

+++ src/test/CircleArea.java
@@ -3,9 +3,9 @@ package test;
 public class CircleArea {
 	//类变量的定义
 	static double PI = 3.14159265;
-	int radius;
+	double radius;
 	public double circumference() {
 		//类变量的使用
-		return 2* PI * this.radius;
+		return 2* PI *this.radius;
 	}
 }
